third_party/blink/web_tests/platform/generic/fast/parser/tabindex-parsing-2-expected.txt

<div tabindex="+4"></div> => 4
<div tabindex="-4"></div> => -4
<div tabindex="+-4"></div> => -1
<div tabindex="-+4"></div> => -1
<div tabindex="+04"></div> => 4
<div tabindex="-04"></div> => -4
<div tabindex="+"></div> => -1
<div tabindex="-"></div> => -1
<div tabindex="    +4"></div> => 4
<div tabindex="    -4"></div> => -4
<div tabindex="    +-4"></div> => -1
<div tabindex="    -+4"></div> => -1
<div tabindex="    +04"></div> => 4
<div tabindex="    -04"></div> => -4
<div tabindex="    +"></div> => -1
<div tabindex="    -"></div> => -1
<div tabindex="+4   "></div> => 4
<div tabindex="-4   "></div> => -4
<div tabindex="+-4   "></div> => -1
<div tabindex="-+4   "></div> => -1
<div tabindex="+04   "></div> => 4
<div tabindex="-04   "></div> => -4
<div tabindex="+    "></div> => -1
<div tabindex="-    "></div> => -1
<div tabindex="&#09;+4"></div> => 4
<div tabindex="&#09;-4"></div> => -4
<div tabindex="&#09;+-4"></div> => -1
<div tabindex="&#09;-+4"></div> => -1
<div tabindex="&#09;+04"></div> => 4
<div tabindex="&#09;-04"></div> => -4
<div tabindex="&#09;+"></div> => -1
<div tabindex="&#09;-"></div> => -1
<div tabindex="+4&#09;"></div> => 4
<div tabindex="-4&#09;"></div> => -4
<div tabindex="+-4&#09;"></div> => -1
<div tabindex="-+4&#09;"></div> => -1
<div tabindex="+04&#09;"></div> => 4
<div tabindex="-04&#09;"></div> => -4
<div tabindex="+&#09;"></div> => -1
<div tabindex="-&#09;"></div> => -1
<div tabindex="
+4"></div> => 4
<div tabindex="
-4"></div> => -4
<div tabindex="
+-4"></div> => -1
<div tabindex="
-+4"></div> => -1
<div tabindex="
+04"></div> => 4
<div tabindex="
-04"></div> => -4
<div tabindex="
+"></div> => -1
<div tabindex="
-"></div> => -1
<div tabindex="+4
"></div> => 4
<div tabindex="-4
"></div> => -4
<div tabindex="+-4
"></div> => -1
<div tabindex="-+4
"></div> => -1
<div tabindex="+04
"></div> => 4
<div tabindex="-04
"></div> => -4
<div tabindex="+
"></div> => -1
<div tabindex="-
"></div> => -1
<div tabindex="&nbsp;+4"></div> => -1
<div tabindex="&nbsp;-4"></div> => -1
<div tabindex="&nbsp;+-4"></div> => -1
<div tabindex="&nbsp;-+4"></div> => -1
<div tabindex="&nbsp;+04"></div> => -1
<div tabindex="&nbsp;-04"></div> => -1
<div tabindex="&nbsp;+"></div> => -1
<div tabindex="&nbsp;-"></div> => -1
<div tabindex="+4&nbsp;"></div> => 4
<div tabindex="-4&nbsp;"></div> => -4
<div tabindex="+-4&nbsp;"></div> => -1
<div tabindex="-+4&nbsp;"></div> => -1
<div tabindex="+04&nbsp;"></div> => 4
<div tabindex="-04&nbsp;"></div> => -4
<div tabindex="+&nbsp;"></div> => -1
<div tabindex="-&nbsp;"></div> => -1
<div tabindex="+4xx"></div> => 4
<div tabindex="-4xx"></div> => -4
<div tabindex="+-4xx"></div> => -1
<div tabindex="-+4xx"></div> => -1
<div tabindex="+04xx"></div> => 4
<div tabindex="-04xx"></div> => -4
<div tabindex="+xx"></div> => -1
<div tabindex="-xx"></div> => -1
<div tabindex="+4"></div> => 4
<div tabindex="-4"></div> => -4
<div tabindex="+-4"></div> => -1
<div tabindex="-+4"></div> => -1
<div tabindex="+04"></div> => 4
<div tabindex="-04"></div> => -4
<div tabindex="+"></div> => -1
<div tabindex="-"></div> => -1
<div tabindex="+4/"></div> => 4
<div tabindex="-4/"></div> => -4
<div tabindex="+-4/"></div> => -1
<div tabindex="-+4/"></div> => -1
<div tabindex="+04/"></div> => 4
<div tabindex="-04/"></div> => -4
<div tabindex="+/"></div> => -1
<div tabindex="-/"></div> => -1
